requirements.txt
asgiref==3.7.2
django-jazzmin==2.6.0
pillow==10.2.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0